Migrate TransactionHistory to shadcn registry format
- Add TransactionHistory component to src/components/w3-kit/
- Extract types and utilities to separate files
- Update page.tsx to use @/components/w3-kit import path
- Update CLI installation to use shadcn registry URL
- Add shadcn configuration files

Closes #31
