+++ src/app/docs/components/transaction-history/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useState, useCallback } from "react";
-import { TransactionHistory } from "./component";
+import { TransactionHistory } from "@/components/w3-kit/transaction-history";
 import { Code, Eye } from "lucide-react";
 import { CodeBlock } from "@/components/docs/codeBlock";
 
@@ -217,7 +217,7 @@ export default function Page() {
                   <p className="text-sm text-gray-600 dark:text-gray-400 mb-4">
                     Run the following command to add the Transaction History component to your project:
                   </p>
-                  <CodeBlock code="npx w3-kit@latest add transaction-history" id="cli" />
+                  <CodeBlock code="npx shadcn@latest add https://w3-kit.com/registry/transaction-history.json" id="cli" />
                   <p className="text-sm text-gray-600 dark:text-gray-400 mt-4">
                     This will:
                   </p>
